+++ CHANGELOG.md
@@ -5,10 +5,15 @@
 ### Added
 
 * config: Added `logging` config field.
+* config: Added `hasura.create_source` flag to create PostgreSQL source if missing.
 
 ### Deprecated
 
 * cli: `--logging-config` option is deprecated.
+
+### Fixed
+
+* hasura: Do not apply table customizations to tables from other sources.
 
 ## 5.1.7 - 2022-06-15
 
@@ -30,7 +35,6 @@
 ### Other
 
 * docs: Migrate docs from GitBook to mdbook.
->>>>>>> master
 
 ## 5.1.5 - 2022-06-05