+++ CHANGELOG.md
@@ -2,13 +2,40 @@
 
 ## [unreleased]
 
+### Added
+
+* config: Added `logging` config field.
+* config: Added `hasura.create_source` flag to create PostgreSQL source if missing.
+
+### Fixed
+
+* hasura: Do not apply table customizations to tables from other sources.
+
 ### Deprecated
 
+* cli: `--logging-config` option is deprecated.
 * cli: All `run` command options are deprecated. Use the `advanced` section of the config.
+
+## 5.1.7 - 2022-06-15
+
+### Fixed
+
+* index: Fixed `token_transfer` index not receiving realtime updates.
+
+## 5.1.6 - 2022-06-08
+
+### Fixed
+
+* cli: Commands with `--help` option no longer require a working DipDup config.
+* index: Fixed crash with `RuntimeError` after continuous realtime connection loss.
 
 ### Performance
 
 * cli: Lazy import dependencies to speed up startup.
+
+### Other
+
+* docs: Migrate docs from GitBook to mdbook.
 
 ## 5.1.5 - 2022-06-05